Added issue_handler moduile, and settings tracker class.
issue_handler contains a set of methods designed to be helper functions for the command line arguments. The purpose of these helper functions is to act as a layer of abstraction between the stored issues, and the CLI.

The tracker class is designed to keep track of non-user set variables such as the increment of the issue count. This class reads its self in from JSON and stores its self into JSON.